samples: move the inboxed handler out of the shared project, and make the sample runnable
Answers the second review on #4331. It found no blocker, but four of its notes were
right and two of those were defects I introduced.

1. [UseInbox] was on a handler in the shared Events project, which all four apps in
   this sample reference and all four scan with AutoFromAssemblies(). Only
   GreetingsReceiverConsole registers an inbox, so the other three registered a
   handler whose pipeline cannot be built — latent, because pipelines are built
   lazily and none of them sends a GreetingEvent in-process, but real. The handler
   now lives in GreetingsReceiverConsole, which owns the policy.

2. My own comment cited ServiceCollectionExtensions.cs:657-666. The branch is :655
   and the inbox arguments are :665/:676, and the filename is ambiguous — there are
   two ServiceCollectionExtensions.cs. It now names AddEventBus, which is stable
   and greppable. A line-number citation that was already wrong when written is
   exactly the defect class this sample exists to document.

3. The committed form was not the form that ran: three files hardcoded a Windows
   SQLEXPRESS connection string, and the Docker verification used an uncommitted
   edit. Both apps now read ConnectionStrings:Brighter and fall back to the
   original SQLEXPRESS default, so Windows readers are unaffected and everyone else
   can point it at a container without editing source. The packages this needs were
   already referenced and unused.

4. GreetingsSender got Microsoft.Extensions.Hosting only transitively via
   BoxProvisioning; it now references it explicitly, since it calls
   Host.CreateApplicationBuilder.

Re-proven end to end against SQL Server 2022 in Docker using the COMMITTED source
and an environment variable, no source edit:

  Provisioned Outbox 'Outbox' successfully
  Decoupled invocation of message: Topic:greeting.event Id:01a08b63-...
  receiver: provisioned Inbox, delivery 1 handler ran 1 time
  receiver: delivery 2 same id, handler ran 0 times
    Command 01a08b63-5db8-711a-905d-9582f4557f77 has already been seen
  InboxMessages rows: 1

Left as follow-ups, per the review: a validation spec comparing a subscription
against its channel factory's type, a test pinning the ChannelFactory downcast,
shared table-name constants, the UTF-16 BrighterSqlQueue.sql, and a README.
